pyproject.toml
[project]
name = "formnow-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
]
